requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.5.1.48
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==1.5
python-decouple==3.4
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
